+++ Statistics/Project.csproj
@@ -2,7 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <StartupObject>ConsoleBasics.StandardNumericFormats</StartupObject>
   </PropertyGroup>
 
 </Project>
